+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 History
 ================================================================================
 
+- 0.7.2
+    - Added: stupidQuestionScriptcode parameter
+    - Added: stupidQuestionNoScript parameter
+    - Documentation: StupidQuestion has to be used as preHook **and** as hook
 - 0.7.1
     - Added: stupidQuestionRegister parameter
 - 0.7